Move buttons to control panel
