+++ src/shared/components/Graph/GraphControlPanel.less
@@ -1,0 +1,18 @@
+.graph-control-panel {
+	position: absolute;
+	left: 0;
+	top: 0;
+	width: 100%;
+	margin-top: 1em;
+	z-index: 10;
+
+	button {
+		margin-right: 4px;
+	}
+
+	.controls {
+		justify-content: space-between;
+		display: flex;
+		width: 100%;
+	}
+}
